review service testing

+++ review-service/src/main/java/com/company/review/dto/admin/ReviewAdminResponse.java
@@ -18,7 +18,7 @@ public class ReviewAdminResponse {
 
     private String comment;
 
-    private Integer rate;
+    private Double rate;
 
     @DateTimeFormat(pattern = "E, dd MMM yyyy HH:mm:ss z")
     private LocalDateTime created;
